mdLinks and CLI working
