+++ READMEmdLinks.md
@@ -1,0 +1,18 @@
+#  👾 MD LINKS 👾
+
+Despúes de instalar la librería. 
+ejecutala con el comando '$ mdLinks' +'ENTER'
+te aparecera un mensaje donde se solicita ingreses una ruta 
+para continuar ejecutar 'ENTER'
+y selecciona una de las opciones:
+
+Validate True: te devolverá un array de objetos con la informacion de cada archivo md de la ruta (name, href, text, status, message Ok or Fail)
+Validate False: te devolverá un array de objetos similar a validate True pero sin el status y el message.
+Stats: te indica el nombre del archivo md que ingresaste o de los que habitan en el directorio que ingresas  con su totalLinks y el total de Links unicos.
+StastAndValidate: te muestra el total de links que hay en la ruta que ingresaste , la ruta que ingresaste, lel total de unicos unicos y un total de links rotos en dicha ruta. 
+
+Puedes explorar las opciones al darle enter despues de ingresar tu ruta. Seleccionas la opcion aprietas enter o delete y te regresa al menu de opciones donde puedes seguir explorando con la ruta ingresada. s
+
+si deseas salir solo ingresa 0 seguido de 'ENTER' 
+
+Node 16
